Add a module dependency check tool.
This calculates dependencies for modules and then tries to link them
against the kernel to see if there are any missing symbols, and might
reveal if you are missing dependencies or spelled something wrong. Due
to the way the current module build system works, this hasn't been
integrated into the Makefile yet.